# Conflicts:
#	Vorb.vcxproj
#	Vorb.vcxproj.filters

+++ include/FastConversion.inl
@@ -21,9 +21,8 @@ public:
         U output = 0;
         // slowest version
 #if defined(__APPLE__) || defined(__linux__) || defined(WT64)
-        // std::floor expects a double
-        output = static_cast<U>(std::floor(static_cast<double>(x)));
-#elif defined(WT32)
+        output = static_cast<U>(std::floor(x));
+#elif defined(OS_WINDOWS) && defined(ARCH_X86_32) && 0 
         __asm {
             fld x;
             fadd st, st(0);
@@ -32,8 +31,7 @@ public:
             sar i, 1;
         };
 #else
-        // std::floor expects a double
-        output = static_cast<U>(std::floor(static_cast<double>(x)));
+        output = static_cast<U>(std::floor(x));
 #endif
         return output;
     }
@@ -45,9 +43,8 @@ public:
     inline static U ceiling(const T& x) {
         U output = 0;
 #if defined(__APPLE__) || defined(__linux__) || defined(WT64)
-        // std::ceil expects a double
-        output = static_cast<U>(std::ceil(static_cast<double>(x)));
-#elif defined(WT32)
+        output = static_cast<U>(std::ceil(x));
+#elif defined(OS_WINDOWS) && defined(ARCH_X86_32) && 0 
         __asm {
             fld x;
             fadd st, st(0);
@@ -56,8 +53,7 @@ public:
             sar i, 1;
         };
 #else
-        // std::ceil expects a double
-        output = static_cast<U>(std::ceil(static_cast<double>(x)));
+        output = static_cast<U>(std::ceil(x));
 #endif
         return output;
     }
@@ -69,14 +65,11 @@ public:
     inline static U trunc(const T& x) {
         U output = 0;
 #if defined(__APPLE__) || defined(__linux__) || defined(WT64)
-        // std::trunc expects a double
-        output = static_cast<U>(std::trunc(static_cast<double>(x)));
-#elif defined(WT32)
-        // std::trunc expects a double
-        output = static_cast<U>(std::trunc(static_cast<double>(x)));
+        output = static_cast<U>(std::trunc(x));
+#elif defined(OS_WINDOWS) && defined(ARCH_X86_32)
+        output = static_cast<U>(std::trunc(x));
 #else
-        // std::trunc expects a double
-        output = static_cast<U>(std::trunc(static_cast<double>(x)));
+        output = static_cast<U>(std::trunc(x));
 #endif
         return output;
     }
@@ -90,14 +83,11 @@ public:
     inline static U round(const T& x) {
         U output = 0;
 #if defined(__APPLE__) || defined(__linux__) || defined(WT64)
-        // std::round expects a double
-        output = static_cast<U>(std::round(static_cast<double>(x)));
-#elif defined(WT32)
-        // std::round expects a double
-        output = static_cast<U>(std::round(static_cast<double>(x)));
+        output = static_cast<U>(std::round(x));
+#elif defined(OS_WINDOWS) && defined(ARCH_X86_32)
+        output = static_cast<U>(std::round(x));
 #else
-        // std::round expects a double
-        output = static_cast<U>(std::round(static_cast<double>(x)));
+        output = static_cast<U>(std::round(x));
 #endif
         return output;
     }
